LaboratorioPrueba/labs/requirements.txt
Django>=5.1.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
requests>=2.31.0  # Para consumir APIs externas